// src/generated/prisma/enums.ts
export const Role = {
  admin: 'admin',
  office: 'office',
  attendance: 'attendance',
  parent: 'parent'
} as const

export type Role = (typeof Role)[keyof typeof Role]